refactor(mouse): move gesture primitives under input::mouse
Relocates the shared click/drag/wheel gesture primitives from
src/input/gesture.rs to src/input/mouse/gesture.rs, since they are
consumed only by mouse code (host button/wheel dispatch and the tmux
gesture). The module is declared pub(in crate::input) for now because
two tmux-mouse importers still live outside input::mouse until Task 5
relocates them; Task 7 narrows visibility to private once that lands.
