+++ src/input/mouse/button.rs
@@ -11,9 +11,9 @@ use super::{
 use crate::input::InputPhase;
 use crate::input::bindings::OzmaMouseConfig;
 use crate::input::current_modifiers;
-use crate::input::gesture::{DragGesture, DragPhase, HeldPointer, OzmaMouseGesture};
 use crate::input::hyperlink::link_modifier_held;
 use crate::input::keyboard::current_terminal_modifiers;
+use crate::input::mouse::gesture::{DragGesture, DragPhase, HeldPointer, OzmaMouseGesture};
 use crate::surface::geometry::topmost_surface_at;
 use bevy::input::ButtonState;
 use bevy::input::mouse::{MouseButton, MouseButtonInput};
